Refactor validate_html to a helper not fixture
- When testing the validate_html fixture, it wasnt being imported. I
think that to implement fixtures that can be used in another library
this requires writing a pytest plugin, however this seemed overly
complicated. Ref https://docs.pytest.org/en/latest/how-to/writing_plugins.html
- Also changed the helper to remove the assert and return the result
of the tidylib check. This allowed the error message to surface in the
project where the helper is used