Remove page param in DSLinkFilter
At the moment, we copy all params from the querystring and edit them based on the filter value. 

This is great for chaining more than one filter, but `page` is a special case: there is no value in deep-linking to page `n` of a newly filtered queryset. If you are adding (or removing) a filter, you will want to see page 1, always.